rustc_attr_parsing

Module attributes

This module defines traits for attribute parsers, little state machines that recognize and parse attributes out of a longer list of attributes. The main trait is called AttributeParser. You can find more docs about AttributeParsers on the trait itself. However, for many types of attributes, implementing AttributeParser is not necessary. It allows for a lot of flexibility you might not want.

Specifically, you might not care about managing the state of your AttributeParser state machine yourself. In this case you can choose to implement:

SingleAttributeParser: makes it easy to implement an attribute which should error if it appears more than once in a list of attributes

CombineAttributeParser: makes it easy to implement an attribute which should combine the contents of attributes, if an attribute appear multiple times in a list

Attributes should be added to crate::context::ATTRIBUTE_PARSERS to be parsed.

Modules

allow_unstable ğŸ”’

cfg ğŸ”’

cfg_old ğŸ”’

codegen_attrs ğŸ”’

confusables ğŸ”’

deprecation ğŸ”’

dummy ğŸ”’

inline ğŸ”’

link_attrs ğŸ”’

lint_helpers ğŸ”’

loop_match ğŸ”’

macro_attrs ğŸ”’

must_use ğŸ”’

no_implicit_prelude ğŸ”’

non_exhaustive ğŸ”’

path ğŸ”’

proc_macro_attrs ğŸ”’

repr ğŸ”’

rustc_internal ğŸ”’

semantics ğŸ”’

stability ğŸ”’

test_attrs ğŸ”’

traits ğŸ”’

transparency ğŸ”’

util ğŸ”’

Structs

Combine ğŸ”’

Use in combination with CombineAttributeParser. Combine<T: CombineAttributeParser> implements AttributeParser.

Single ğŸ”’

Use in combination with SingleAttributeParser. Single<T: SingleAttributeParser> implements AttributeParser.

WithoutArgs ğŸ”’

Enums

AttributeOrder ğŸ”’

OnDuplicate ğŸ”’

Traits

AttributeParser ğŸ”’

An AttributeParser is a type which searches for syntactic attributes.

CombineAttributeParser ğŸ”’

Alternative to AttributeParser that automatically handles state management. If multiple attributes appear on an element, combines the values of each into a ThinVec. Combine<T> where T: CombineAttributeParser implements AttributeParser.

NoArgsAttributeParser ğŸ”’

An even simpler version of SingleAttributeParser: now automatically check that there are no arguments provided to the attribute.

SingleAttributeParser ğŸ”’

Alternative to AttributeParser that automatically handles state management. A slightly simpler and more restricted way to convert attributes. Assumes that an attribute can only appear a single time on an item, and errors when it sees more.

Type Aliases

AcceptFn ğŸ”’

AcceptMapping ğŸ”’

ConvertFn ğŸ”’