rustc_attr_parsing

Module context

All the important types given to attribute parsers when parsing

Modules

early 🔒

late 🔒

private 🔒

Macros

attribute_parsers 🔒

Structs

AcceptContext

Context given to every attribute parser when accepting

Early

used when parsing attributes for miscellaneous things before ast lowering

FinalizeContext 🔒

Context given to every attribute parser during finalization.

GroupTypeInner 🔒

GroupTypeInnerAccept 🔒

Late

used when parsing attributes during ast lowering

SharedContext

Context given to every attribute parser during finalization.

Enums

OmitDoc

ShouldEmit

Traits

Stage

Type Aliases

AcceptFn 🔒

FinalizeFn 🔒

GroupType 🔒