rustc_attr_parsing::context::late

Type Alias Combine

type Combine<T> = Combine<T, Late>;

Aliased Type

struct Combine<T> {
    phantom: PhantomData<(Late, T)>,
    items: ThinVec<<T as CombineAttributeParser<Late>>::Item>,
    first_span: Option<Span>,
}

Fields
phantom: PhantomData<(Late, T)>items: ThinVec<<T as CombineAttributeParser<Late>>::Item>

A list of all items produced by parsing attributes so far. One attribute can produce any amount of items.

first_span: Option<Span>

The full span of the first attribute that was encountered.

Layout

Note: Encountered an error during type layout; the type failed to be normalized.