rustc_attr_parsing::context

Struct AcceptContext

pub struct AcceptContext<'f, 'sess, S: Stage> {
    pub(crate) shared: SharedContext<'f, 'sess, S>,
    pub(crate) attr_span: Span,
    pub(crate) template: &'f AttributeTemplate,
    pub(crate) attr_path: AttrPath,
}

Context given to every attribute parser when accepting

Gives AttributeParsers enough information to create errors, for example.

Fields
shared: SharedContext<'f, 'sess, S>attr_span: Span

The span of the attribute currently being parsed

template: &'f AttributeTemplate

The expected structure of the attribute.

Used in reporting errors to give a hint to users what the attribute should look like.

attr_path: AttrPath

The name of the attribute we’re currently accepting.

Implementations

impl<'f, 'sess: 'f, S: Stage> AcceptContext<'f, 'sess, S>

pub(crate) fn unknown_key( &self, span: Span, found: String, options: &'static [&'static str], ) -> ErrorGuaranteed

pub(crate) fn expected_string_literal( &self, span: Span, actual_literal: Option<&MetaItemLit>, ) -> ErrorGuaranteed

error that a string literal was expected. You can optionally give the literal you did find (which you found not to be a string literal) which can make better errors. For example, if the literal was a byte string it will suggest removing the b prefix.

pub(crate) fn expected_integer_literal(&self, span: Span) -> ErrorGuaranteed

pub(crate) fn expected_list(&self, span: Span) -> ErrorGuaranteed

pub(crate) fn expected_no_args(&self, args_span: Span) -> ErrorGuaranteed

pub(crate) fn expected_identifier(&self, span: Span) -> ErrorGuaranteed

emit an error that a name was expected here

pub(crate) fn expected_name_value( &self, span: Span, name: Option<Symbol>, ) -> ErrorGuaranteed

emit an error that a name = value pair was expected at this span. The symbol can be given for a nicer error message talking about the specific name that was found lacking a value.

pub(crate) fn duplicate_key(&self, span: Span, key: Symbol) -> ErrorGuaranteed

emit an error that a name = value pair was found where that name was already seen.

pub(crate) fn unexpected_literal(&self, span: Span) -> ErrorGuaranteed

an error that should be emitted when a MetaItemOrLitParser was expected not to be a literal, but instead a meta item.

pub(crate) fn expected_single_argument(&self, span: Span) -> ErrorGuaranteed

pub(crate) fn expected_at_least_one_argument( &self, span: Span, ) -> ErrorGuaranteed

pub(crate) fn expected_specific_argument( &self, span: Span, possibilities: Vec<&'static str>, ) -> ErrorGuaranteed

pub(crate) fn expected_specific_argument_and_list( &self, span: Span, possibilities: Vec<&'static str>, ) -> ErrorGuaranteed

pub(crate) fn expected_specific_argument_strings( &self, span: Span, possibilities: Vec<&'static str>, ) -> ErrorGuaranteed

pub(crate) fn warn_empty_attribute(&mut self, span: Span)