rustc_attr_parsing::context

Struct AttributeParser

pub struct AttributeParser<'sess, S: Stage = Late> {
    pub(crate) tools: Vec<Symbol>,
    features: Option<&'sess Features>,
    sess: &'sess Session,
    stage: S,
    parse_only: Option<Symbol>,
}

Context created once, for example as part of the ast lowering context, through which all attributes can be lowered.

Fields
tools: Vec<Symbol>features: Option<&'sess Features>sess: &'sess Sessionstage: Sparse_only: Option<Symbol>

Only parse attributes with this symbol.

Used in cases where we want the lowering infrastructure for parse just a single attribute.

Implementations

impl<'sess> AttributeParser<'sess, Early>

pub fn parse_limited( sess: &'sess Session, attrs: &[Attribute], sym: Symbol, target_span: Span, target_node_id: NodeId, features: Option<&'sess Features>, ) -> Option<Attribute>

This method allows you to parse attributes before you have access to features or tools. One example where this is necessary, is to parse feature attributes themselves for example.

Try to use this as little as possible. Attributes should be lowered during rustc_ast_lowering. Some attributes require access to features to parse, which would crash if you tried to do so through parse_limited.

To make sure use is limited, supply a Symbol you’d like to parse. Only attributes with that symbol are picked out of the list of instructions and parsed. Those are returned.

No diagnostics will be emitted when parsing limited. Lints are not emitted at all, while errors will be emitted as a delayed bugs. in other words, we expect attributes parsed with parse_limited to be reparsed later during ast lowering where we do emit the errors

pub fn parse_single<T>( sess: &'sess Session, attr: &Attribute, target_span: Span, target_node_id: NodeId, features: Option<&'sess Features>, emit_errors: ShouldEmit, parse_fn: fn(cx: &mut AcceptContext<'_, '_, Early>, item: &ArgParser<'_>) -> T, template: &AttributeTemplate, ) -> T

impl<'sess, S: Stage> AttributeParser<'sess, S>

pub fn new( sess: &'sess Session, features: &'sess Features, tools: Vec<Symbol>, stage: S, ) -> Self

pub(crate) fn sess(&self) -> &'sess Session

pub(crate) fn features(&self) -> &'sess Features

pub(crate) fn features_option(&self) -> Option<&'sess Features>

pub(crate) fn dcx(&self) -> DiagCtxtHandle<'sess>

pub fn parse_attribute_list( &mut self, attrs: &[Attribute], target_span: Span, target_id: S::Id, omit_doc: OmitDoc, lower_span: impl Copy + Fn(Span) -> Span, emit_lint: impl FnMut(AttributeLint<S::Id>), ) -> Vec<Attribute>

Parse a list of attributes.

target_span is the span of the thing this list of attributes is applied to, and when omit_doc is set, doc attributes are filtered out.

pub fn is_parsed_attribute(path: &[Symbol]) -> bool

Returns whether there is a parser for an attribute with this name

fn lower_attr_args( &self, args: &AttrArgs, lower_span: impl Fn(Span) -> Span, ) -> AttrArgs