rustc_attr_parsing::context

Struct SharedContext

pub struct SharedContext<'p, 'sess, S: Stage> {
    pub(crate) cx: &'p mut AttributeParser<'sess, S>,
    pub(crate) target_span: Span,
    pub(crate) target_id: S::Id,
    emit_lint: &'p mut dyn FnMut(AttributeLint<S::Id>),
}

Context given to every attribute parser during finalization.

Gives AttributeParsers enough information to create errors, for example.

Fields
cx: &'p mut AttributeParser<'sess, S>

The parse context, gives access to the session and the diagnostics context.

target_span: Span

The span of the syntactical component this attribute was applied to

target_id: S::Id

The id (NodeId if S is Early, HirId if S is Late) of the syntactical component this attribute was applied to

emit_lint: &'p mut dyn FnMut(AttributeLint<S::Id>)
Implementations

impl<'f, 'sess: 'f, S: Stage> SharedContext<'f, 'sess, S>

pub(crate) fn emit_err( &self, diag: impl for<'x> Diagnostic<'x>, ) -> ErrorGuaranteed

pub(crate) fn emit_lint(&mut self, lint: AttributeLintKind, span: Span)

Emit a lint. This method is somewhat special, since lints emitted during attribute parsing must be delayed until after HIR is built. This method will take care of the details of that.

pub(crate) fn warn_unused_duplicate( &mut self, used_span: Span, unused_span: Span, )

pub(crate) fn warn_unused_duplicate_future_error( &mut self, used_span: Span, unused_span: Span, )