rustc_attr_parsing

Module parser

This is in essence an (improved) duplicate of rustc_ast/attr/mod.rs. That module is intended to be deleted in its entirety.

FIXME(jdonszelmann): delete rustc_ast/attr/mod.rs

Structs

MetaItemListParser

MetaItemListParserContext 🔒

MetaItemParser

Utility that deconstructs a MetaItem into usable parts.

NameValueParser

SegmentIterator

Enums

ArgParser

MetaItemOrLitParser

Inside lists, values could be either literals, or more deeply nested meta items. This enum represents that.

PathParser

Functions

expr_to_lit 🔒